More code, less macros.  Less raw pointers!
Macros:
Replaced some by typedef, removed others entirely.

Pointers:
I Removed the need for a "IVisitor*" as argument in the node->accept methods. Instead, it takes a reference, that avoid the user to declare a, for instance, "new Dumper()" to pass as argument. Instead, he can just pass Dumper(). It avoids memory leaks and pointer manipulation !

The last place where I use raw pointers is in the Visitor's visit method, but they're always passed "this" objects. So, I think it's fine.