TypeChecker rework !
It looks like this is the best solution : cleaner, easier to read and more stable, but I'll test it more, and review the code a few times before pushing it on the main branch.

After i'm done implementing Callables (variables) as a derived of ASTValue and the assignement operator is implemented i'll be able to finish the typechecker completly.