Operation Health : Part I
I'm currently making some relatively big changes to how the code works currently.

I've changed the var:var_attr, it's no longer located in the FVal, varRef replaced it.

I'm now using more enum classes for clarity, and made some changes to the behaviour of the Symbols table : it will now attempt to implicitely cast the value into the desired type is there is a mismatch between the type of the FVal passed to setEntry and the type of the variable you're trying to assign to !

Next step is changing the error reporting system to use the context class, and use exceptions for critical errors.