Lexer Overhaul & Bugfixing
I took a look at my lexer and found several bugs, one included the escape sequences detection. I've reworked it, and it's cleaner and clearer now, but it's still far from perfect, but it does the job really well ! Why overcomplicate it if it isn't needed !

I also did some profiling for fun, and found that regex_match was causing a (relatively low, but still a) performance problem, so I've deleted it and I now use stringstreams. Solved it ! Now most of the time is spent on the parsing and that's normal.

All of the tests run under 55ms, it's correct for now and I can't do much more until the project gets more complex. Also, I'll avoid wasting time optimizing stuff that doesn't need it. "Premature optimization is the root of all evil!"