Changed error messages.
They're still far from perfect and have ZERO awareness of the current context. Making user-friendly error messages will need a complete parser refactoring, but that is really not a priority. It will be done, but I prefer to finish my language first, then see if it's really needed. Maybe just adding a column infomation or a prettier presentation might help since the syntax is trivial?