String manipulators are not polymorphic anymore.
It's just too much of a mess. Every single std::string in the interpreter will be considered encoded as UTF-8.

Also, it lets me separate the context for UTF8StringManipulator (god that's a long name) which lets other functions that doesn't have access to the current context use it freely.