Implemented ParsingResult system.
It's far from perfect, and most of the time the parser doesn't take advantage of it. In the coming days, i'll do a relecture of my code and modify as much code as possible to take advantage of this new system, and also add a call to die() whenever an error is found (stop the parsing).

After that, I'll rush the rest of the parser (functions) so I can finally rest and bit an start the work on semantic analysis before implementing support for packages.