Changed Type* -> const Type*
As types derived are immutable once created, all of their member functions are marked const. Making the pointers const doesn't change anything and might enable some optimisations on the compiler's side, so I think this doesn't hurt and it's a good practice. It shows clearly that types are immutable when reading the code.