Simplifying the Constraints hierarchy.
It's now only 1 class instead of a hierarchy, and an enum in class tells you what's the kind of the constraint.

This is done because I can't think of any situation where i'll need a constraint that doesn't fit in a single object. I just don't need a hierarchy for this!