[SemaExpr] Forbidding array literals of function types and improving diagnostics when one element is an ErrorType
About the FunctionType thing, it's done because Functions aren't first class citizen in Fox, so this doesn't make sense. In the future, they might be first class, but not for now.

About the improved diags: for instance we had an inference error for the first [] in this example: [[],[1],[some erroneous expr]]. Now there is no such error because the typechecking function will use the [1] to infer that [] is an [int].